% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_dom_element.R
\name{download_dom_element}
\alias{download_dom_element}
\title{Download a DOM element in a Shiny app to a PNG}
\usage{
download_dom_element(
  id,
  name,
  session = shiny::getDefaultReactiveDomain(),
  asis = FALSE
)
}
\arguments{
\item{id}{Shiny input ID of element to take a snapshot of}

\item{name}{Name of output file (without extension)}
}
\description{
Download a DOM element in a Shiny app to a PNG
}
