% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toggle_leaflet.R
\name{hide_leaflet_zoom}
\alias{hide_leaflet_zoom}
\alias{show_leaflet_zoom}
\title{Hide or show leaflet control buttons}
\usage{
hide_leaflet_zoom(session = shiny::getDefaultReactiveDomain())

show_leaflet_zoom(session = shiny::getDefaultReactiveDomain())
}
\description{
When saving a leaflet map, use these functions to first remove (and then add) the control
buttons, so they do not appear in the screenshot
}
