% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_dom_element.R
\name{save_dom_element}
\alias{save_dom_element}
\title{Save a DOM element in a Shiny app to a PNG}
\usage{
save_dom_element(id, id_out, session = getDefaultReactiveDomain())
}
\arguments{
\item{id}{Shiny input ID of element to take a snapshot of}

\item{id_out}{Snapshot will become available as this input ID (e.g. input$myelement)}
}
\description{
Save a DOM element in a Shiny app to a PNG
}
