% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_dom_element_dependencies.R
\name{save_dom_element_dependencies}
\alias{save_dom_element_dependencies}
\title{Save DOM element shiny dependencies}
\usage{
save_dom_element_dependencies(tag)
}
\description{
Save DOM element shiny dependencies
}
